+++ source/server/watchdog_impl.h
@@ -19,13 +19,12 @@ public:
   /**
    * @param interval WatchDog timer interval (used after startWatchdog())
    */
-  WatchDogImpl(Thread::ThreadIdPtr&& thread_id, TimeSource& tsource,
-               std::chrono::milliseconds interval)
-      : thread_id_(std::move(thread_id)), time_source_(tsource),
+  WatchDogImpl(Thread::ThreadId thread_id, TimeSource& tsource, std::chrono::milliseconds interval)
+      : thread_id_(thread_id), time_source_(tsource),
         latest_touch_time_since_epoch_(tsource.monotonicTime().time_since_epoch()),
         timer_interval_(interval) {}
 
-  const Thread::ThreadId& threadId() const override { return *thread_id_; }
+  Thread::ThreadId threadId() const override { return thread_id_; }
   MonotonicTime lastTouchTime() const override {
     return MonotonicTime(latest_touch_time_since_epoch_.load());
   }
@@ -37,7 +36,7 @@ public:
   }
 
 private:
-  Thread::ThreadIdPtr thread_id_;
+  const Thread::ThreadId thread_id_;
   TimeSource& time_source_;
   std::atomic<std::chrono::steady_clock::duration> latest_touch_time_since_epoch_;
   Event::TimerPtr timer_;