+++ test/common/router/router_test.cc
@@ -48,6 +48,7 @@ using testing::Return;
 using testing::ReturnPointee;
 using testing::ReturnRef;
 using testing::SaveArg;
+using testing::StartsWith;
 
 namespace Envoy {
 namespace Router {
@@ -79,11 +80,12 @@ public:
 
 class RouterTestBase : public testing::Test {
 public:
-  RouterTestBase(bool start_child_span, bool suppress_envoy_headers)
+  RouterTestBase(bool start_child_span, bool suppress_envoy_headers,
+                 Protobuf::RepeatedPtrField<std::string> strict_headers_to_check)
       : http_context_(stats_store_.symbolTable()), shadow_writer_(new MockShadowWriter()),
         config_("test.", local_info_, stats_store_, cm_, runtime_, random_,
                 ShadowWriterPtr{shadow_writer_}, true, start_child_span, suppress_envoy_headers,
-                test_time_.timeSystem(), http_context_),
+                std::move(strict_headers_to_check), test_time_.timeSystem(), http_context_),
         router_(config_) {
     router_.setDecoderFilterCallbacks(callbacks_);
     upstream_locality_.set_zone("to_az");
@@ -256,14 +258,15 @@ public:
 
 class RouterTest : public RouterTestBase {
 public:
-  RouterTest() : RouterTestBase(false, false) {
+  RouterTest() : RouterTestBase(false, false, Protobuf::RepeatedPtrField<std::string>{}) {
     EXPECT_CALL(callbacks_, activeSpan()).WillRepeatedly(ReturnRef(span_));
   };
 };
 
 class RouterTestSuppressEnvoyHeaders : public RouterTestBase {
 public:
-  RouterTestSuppressEnvoyHeaders() : RouterTestBase(false, true) {}
+  RouterTestSuppressEnvoyHeaders()
+      : RouterTestBase(false, true, Protobuf::RepeatedPtrField<std::string>{}) {}
 };
 
 TEST_F(RouterTest, RouteNotFound) {
@@ -1092,7 +1095,12 @@ TEST_F(RouterTest, ResetDuringEncodeHeaders) {
 
   Http::TestHeaderMapImpl headers;
   HttpTestUtility::addDefaultHeaders(headers);
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
+  // First connection is successful and reset happens later on.
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _));
   router_.decodeHeaders(headers, true);
   EXPECT_TRUE(verifyHostUpstreamStats(0, 1));
 }
@@ -1128,7 +1136,8 @@ TEST_F(RouterTest, UpstreamTimeout) {
   EXPECT_CALL(callbacks_, encodeHeaders_(HeaderMapEqualRef(&response_headers), false));
   EXPECT_CALL(callbacks_, encodeData(_, true));
   EXPECT_CALL(*router_.retry_state_, shouldRetryReset(_, _)).Times(0);
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, _));
   response_timeout_->callback_();
 
   EXPECT_EQ(1U,
@@ -1259,7 +1268,8 @@ TEST_F(RouterTest, GrpcReset) {
   EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(200));
   response_decoder->decodeHeaders(std::move(response_headers), false);
   EXPECT_TRUE(verifyHostUpstreamStats(0, 0));
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _));
   encoder1.stream_.resetStream(Http::StreamResetReason::RemoteReset);
   EXPECT_TRUE(verifyHostUpstreamStats(0, 1));
   EXPECT_EQ(1UL, stats_store_.counter("test.rq_reset_after_downstream_response_started").value());
@@ -1347,7 +1357,9 @@ TEST_F(RouterTest, UpstreamTimeoutWithAltResponse) {
   Http::TestHeaderMapImpl response_headers{{":status", "204"}};
   EXPECT_CALL(callbacks_, encodeHeaders_(HeaderMapEqualRef(&response_headers), true));
   EXPECT_CALL(*router_.retry_state_, shouldRetryReset(_, _)).Times(0);
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(204));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(204)));
   response_timeout_->callback_();
 
   EXPECT_EQ(1U,
@@ -1393,7 +1405,9 @@ TEST_F(RouterTest, UpstreamPerTryTimeout) {
       {":status", "504"}, {"content-length", "24"}, {"content-type", "text/plain"}};
   EXPECT_CALL(callbacks_, encodeHeaders_(HeaderMapEqualRef(&response_headers), false));
   EXPECT_CALL(callbacks_, encodeData(_, true));
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
   per_try_timeout_->callback_();
 
   EXPECT_EQ(1U, cm_.thread_local_cluster_.cluster_.info_->stats_store_
@@ -1443,7 +1457,8 @@ TEST_F(RouterTest, UpstreamPerTryTimeoutDelayedPoolReady) {
       {":status", "504"}, {"content-length", "24"}, {"content-type", "text/plain"}};
   EXPECT_CALL(callbacks_, encodeHeaders_(HeaderMapEqualRef(&response_headers), false));
   EXPECT_CALL(callbacks_, encodeData(_, true));
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, _));
   per_try_timeout_->callback_();
 
   EXPECT_EQ(1U, cm_.thread_local_cluster_.cluster_.info_->stats_store_
@@ -1489,7 +1504,8 @@ TEST_F(RouterTest, UpstreamPerTryTimeoutExcludesNewStream) {
   pool_callbacks->onPoolReady(encoder, cm_.conn_pool_.host_);
 
   EXPECT_CALL(encoder.stream_, resetStream(Http::StreamResetReason::LocalReset));
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, _));
   EXPECT_CALL(*per_try_timeout_, disableTimer());
   EXPECT_CALL(*response_timeout_, disableTimer());
   EXPECT_CALL(callbacks_.stream_info_,
@@ -1523,6 +1539,10 @@ TEST_F(RouterTest, HedgedPerTryTimeoutFirstRequestSucceeds) {
         callbacks.onPoolReady(encoder1, cm_.conn_pool_.host_);
         return nullptr;
       }));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)))
+      .Times(2);
   expectPerTryTimerCreate();
   expectResponseTimerCreate();
 
@@ -1530,9 +1550,10 @@ TEST_F(RouterTest, HedgedPerTryTimeoutFirstRequestSucceeds) {
   HttpTestUtility::addDefaultHeaders(headers);
   router_.decodeHeaders(headers, true);
 
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
   EXPECT_CALL(encoder1.stream_, resetStream(_)).Times(0);
-
   NiceMock<Http::MockStreamEncoder> encoder2;
   Http::StreamDecoder* response_decoder2 = nullptr;
   router_.retry_state_->expectHedgedPerTryTimeoutRetry();
@@ -1596,6 +1617,11 @@ TEST_F(RouterTest, HedgedPerTryTimeoutThirdRequestSucceeds) {
   EXPECT_CALL(encoder1.stream_, resetStream(_)).Times(0);
 
   Http::HeaderMapPtr response_headers1(new Http::TestHeaderMapImpl{{":status", "500"}});
+  // Local origin connect success happens for first and third try.
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)))
+      .Times(2);
   EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(500));
   EXPECT_CALL(encoder1.stream_, resetStream(_)).Times(0);
   EXPECT_CALL(callbacks_, encodeHeaders_(_, _)).Times(0);
@@ -1619,7 +1645,9 @@ TEST_F(RouterTest, HedgedPerTryTimeoutThirdRequestSucceeds) {
 
   // Now trigger a per try timeout on the 2nd request, expect a 3rd
   router_.retry_state_->expectHedgedPerTryTimeoutRetry();
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
   NiceMock<Http::MockStreamEncoder> encoder3;
   Http::StreamDecoder* response_decoder3 = nullptr;
   EXPECT_CALL(cm_.conn_pool_, newStream(_, _))
@@ -1672,6 +1700,10 @@ TEST_F(RouterTest, RetryOnlyOnceForSameUpstreamRequest) {
         callbacks.onPoolReady(encoder1, cm_.conn_pool_.host_);
         return nullptr;
       }));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)))
+      .Times(2);
   expectPerTryTimerCreate();
   expectResponseTimerCreate();
 
@@ -1681,7 +1713,9 @@ TEST_F(RouterTest, RetryOnlyOnceForSameUpstreamRequest) {
 
   EXPECT_CALL(encoder1.stream_, resetStream(_)).Times(0);
 
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
   router_.retry_state_->expectHedgedPerTryTimeoutRetry();
   per_try_timeout_->callback_();
 
@@ -1706,7 +1740,9 @@ TEST_F(RouterTest, RetryOnlyOnceForSameUpstreamRequest) {
   EXPECT_CALL(*router_.retry_state_, wouldRetryFromHeaders(_)).WillOnce(Return(true));
   response_decoder1->decodeHeaders(std::move(response_headers1), true);
 
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
 
   response_timeout_->callback_();
 }
@@ -1728,6 +1764,10 @@ TEST_F(RouterTest, BadHeadersDroppedIfPreviousRetryScheduled) {
         callbacks.onPoolReady(encoder1, cm_.conn_pool_.host_);
         return nullptr;
       }));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)))
+      .Times(2);
   expectPerTryTimerCreate();
   expectResponseTimerCreate();
 
@@ -1737,7 +1777,9 @@ TEST_F(RouterTest, BadHeadersDroppedIfPreviousRetryScheduled) {
 
   EXPECT_CALL(encoder1.stream_, resetStream(_)).Times(0);
 
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
   router_.retry_state_->expectHedgedPerTryTimeoutRetry();
   per_try_timeout_->callback_();
 
@@ -1798,7 +1840,8 @@ TEST_F(RouterTest, RetryRequestNotComplete) {
   router_.decodeHeaders(headers, false);
 
   router_.retry_state_->expectResetRetry();
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _));
   encoder1.stream_.resetStream(Http::StreamResetReason::RemoteReset);
   EXPECT_TRUE(verifyHostUpstreamStats(0, 1));
 }
@@ -1818,6 +1861,10 @@ TEST_F(RouterTest, HedgedPerTryTimeoutGlobalTimeout) {
         callbacks.onPoolReady(encoder1, cm_.conn_pool_.host_);
         return nullptr;
       }));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)))
+      .Times(2);
   expectPerTryTimerCreate();
   expectResponseTimerCreate();
 
@@ -1825,7 +1872,9 @@ TEST_F(RouterTest, HedgedPerTryTimeoutGlobalTimeout) {
   HttpTestUtility::addDefaultHeaders(headers);
   router_.decodeHeaders(headers, true);
 
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
   EXPECT_CALL(encoder1.stream_, resetStream(_)).Times(0);
   EXPECT_CALL(callbacks_, encodeHeaders_(_, _)).Times(0);
   router_.retry_state_->expectHedgedPerTryTimeoutRetry();
@@ -1849,7 +1898,9 @@ TEST_F(RouterTest, HedgedPerTryTimeoutGlobalTimeout) {
   // Now trigger global timeout, expect everything to be reset
   EXPECT_CALL(encoder1.stream_, resetStream(_)).Times(1);
   EXPECT_CALL(encoder2.stream_, resetStream(_)).Times(1);
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
 
   EXPECT_CALL(callbacks_, encodeHeaders_(_, _))
       .WillOnce(Invoke([&](Http::HeaderMap& headers, bool) -> void {
@@ -1876,6 +1927,10 @@ TEST_F(RouterTest, HedgingRetriesExhaustedBadResponse) {
         callbacks.onPoolReady(encoder1, cm_.conn_pool_.host_);
         return nullptr;
       }));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)))
+      .Times(1);
   expectPerTryTimerCreate();
   expectResponseTimerCreate();
 
@@ -1883,7 +1938,9 @@ TEST_F(RouterTest, HedgingRetriesExhaustedBadResponse) {
   HttpTestUtility::addDefaultHeaders(headers);
   router_.decodeHeaders(headers, true);
 
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
   EXPECT_CALL(encoder1.stream_, resetStream(_)).Times(0);
   EXPECT_CALL(callbacks_, encodeHeaders_(_, _)).Times(0);
   router_.retry_state_->expectHedgedPerTryTimeoutRetry();
@@ -1899,6 +1956,10 @@ TEST_F(RouterTest, HedgingRetriesExhaustedBadResponse) {
         callbacks.onPoolReady(encoder2, cm_.conn_pool_.host_);
         return nullptr;
       }));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)))
+      .Times(1);
   expectPerTryTimerCreate();
   router_.retry_state_->callback_();
 
@@ -1946,6 +2007,14 @@ TEST_F(RouterTest, HedgingRetriesProceedAfterReset) {
         callbacks.onPoolReady(encoder1, cm_.conn_pool_.host_);
         return nullptr;
       }));
+  // First is reset
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _))
+      .Times(1);
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)))
+      .Times(2);
   expectPerTryTimerCreate();
   expectResponseTimerCreate();
 
@@ -1953,7 +2022,9 @@ TEST_F(RouterTest, HedgingRetriesProceedAfterReset) {
   HttpTestUtility::addDefaultHeaders(headers);
   router_.decodeHeaders(headers, true);
 
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
   EXPECT_CALL(encoder1.stream_, resetStream(_)).Times(0);
   EXPECT_CALL(callbacks_, encodeHeaders_(_, _)).Times(0);
   router_.retry_state_->expectHedgedPerTryTimeoutRetry();
@@ -1975,7 +2046,6 @@ TEST_F(RouterTest, HedgingRetriesProceedAfterReset) {
   EXPECT_TRUE(verifyHostUpstreamStats(0, 0));
 
   // Now trigger an upstream reset in response to the first request.
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
   EXPECT_CALL(encoder1.stream_, resetStream(_));
   encoder1.stream_.resetStream(Http::StreamResetReason::RemoteReset);
 
@@ -2016,6 +2086,10 @@ TEST_F(RouterTest, HedgingRetryImmediatelyReset) {
         callbacks.onPoolReady(encoder, cm_.conn_pool_.host_);
         return nullptr;
       }));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                        absl::optional<uint64_t>(absl::nullopt)))
+      .Times(1);
 
   Http::TestHeaderMapImpl headers{{"x-envoy-upstream-rq-per-try-timeout-ms", "5"}};
   HttpTestUtility::addDefaultHeaders(headers);
@@ -2029,7 +2103,9 @@ TEST_F(RouterTest, HedgingRetryImmediatelyReset) {
   router_.retry_state_->expectHedgedPerTryTimeoutRetry();
   EXPECT_EQ(Http::FilterDataStatus::StopIterationNoBuffer, router_.decodeData(*body_data, true));
 
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(
+      cm_.conn_pool_.host_->outlier_detector_,
+      putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, absl::optional<uint64_t>(504)));
   EXPECT_CALL(encoder.stream_, resetStream(_)).Times(0);
   EXPECT_CALL(callbacks_, encodeHeaders_(_, _)).Times(0);
   per_try_timeout_->callback_();
@@ -2039,7 +2115,8 @@ TEST_F(RouterTest, HedgingRetryImmediatelyReset) {
       .WillOnce(Invoke([&](Http::StreamDecoder&, Http::ConnectionPool::Callbacks& callbacks)
                            -> Http::ConnectionPool::Cancellable* {
         EXPECT_CALL(*router_.retry_state_, onHostAttempted(_));
-        EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
+        EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+                    putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _));
         callbacks.onPoolFailure(Http::ConnectionPool::PoolFailureReason::ConnectionFailure,
                                 absl::string_view(), cm_.conn_pool_.host_);
         return nullptr;
@@ -2089,7 +2166,8 @@ TEST_F(RouterTest, RetryNoneHealthy) {
   router_.decodeHeaders(headers, true);
 
   router_.retry_state_->expectResetRetry();
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _));
   encoder1.stream_.resetStream(Http::StreamResetReason::LocalReset);
 
   EXPECT_CALL(cm_, httpConnPoolForCluster(_, _, _, _)).WillOnce(Return(nullptr));
@@ -2120,7 +2198,8 @@ TEST_F(RouterTest, RetryUpstreamReset) {
   router_.decodeHeaders(headers, true);
 
   router_.retry_state_->expectResetRetry();
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _));
   encoder1.stream_.resetStream(Http::StreamResetReason::RemoteReset);
 
   // We expect this reset to kick off a new request.
@@ -2129,6 +2208,9 @@ TEST_F(RouterTest, RetryUpstreamReset) {
       .WillOnce(Invoke([&](Http::StreamDecoder& decoder, Http::ConnectionPool::Callbacks& callbacks)
                            -> Http::ConnectionPool::Cancellable* {
         response_decoder = &decoder;
+        EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+                    putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                              absl::optional<uint64_t>(absl::nullopt)));
         callbacks.onPoolReady(encoder2, cm_.conn_pool_.host_);
         return nullptr;
       }));
@@ -2167,7 +2249,8 @@ TEST_F(RouterTest, RetryUpstreamPerTryTimeout) {
   router_.decodeHeaders(headers, true);
 
   router_.retry_state_->expectResetRetry();
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(504));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_TIMEOUT, _));
   per_try_timeout_->callback_();
   EXPECT_TRUE(verifyHostUpstreamStats(0, 1));
 
@@ -2178,6 +2261,9 @@ TEST_F(RouterTest, RetryUpstreamPerTryTimeout) {
       .WillOnce(Invoke([&](Http::StreamDecoder& decoder, Http::ConnectionPool::Callbacks& callbacks)
                            -> Http::ConnectionPool::Cancellable* {
         response_decoder = &decoder;
+        EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+                    putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_SUCCESS,
+                              absl::optional<uint64_t>(absl::nullopt)));
         callbacks.onPoolReady(encoder2, cm_.conn_pool_.host_);
         return nullptr;
       }));
@@ -2294,7 +2380,8 @@ TEST_F(RouterTest, RetryUpstreamResetResponseStarted) {
   response_decoder->decodeHeaders(std::move(response_headers), false);
   absl::string_view rc_details2 = "upstream_reset_after_response_started{remote reset}";
   EXPECT_CALL(callbacks_.stream_info_, setResponseCodeDetails(rc_details2));
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _));
   encoder1.stream_.resetStream(Http::StreamResetReason::RemoteReset);
   // For normal HTTP, once we have a 200 we consider this a success, even if a
   // later reset occurs.
@@ -2327,7 +2414,8 @@ TEST_F(RouterTest, RetryUpstreamReset100ContinueResponseStarted) {
   EXPECT_EQ(
       1U,
       cm_.thread_local_cluster_.cluster_.info_->stats_store_.counter("upstream_rq_100").value());
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _));
   encoder1.stream_.resetStream(Http::StreamResetReason::RemoteReset);
 }
 
@@ -4006,14 +4094,15 @@ TEST_F(WatermarkTest, RetryRequestNotComplete) {
   router_.decodeData(data, false);
 
   // This should not trigger a retry as the retry state has been deleted.
-  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_, putHttpResponseCode(503));
+  EXPECT_CALL(cm_.conn_pool_.host_->outlier_detector_,
+              putResult(Upstream::Outlier::Result::LOCAL_ORIGIN_CONNECT_FAILED, _));
   encoder1.stream_.resetStream(Http::StreamResetReason::RemoteReset);
   EXPECT_EQ(callbacks_.details_, "upstream_reset_before_response_started{remote reset}");
 }
 
 class RouterTestChildSpan : public RouterTestBase {
 public:
-  RouterTestChildSpan() : RouterTestBase(true, false) {}
+  RouterTestChildSpan() : RouterTestBase(true, false, Protobuf::RepeatedPtrField<std::string>{}) {}
 };
 
 // Make sure child spans start/inject/finish with a normal flow.
@@ -4080,6 +4169,169 @@ TEST_F(RouterTestChildSpan, ResetFlow) {
 
   EXPECT_CALL(*child_span, finishSpan());
   encoder.stream_.resetStream(Http::StreamResetReason::RemoteReset);
+}
+
+Protobuf::RepeatedPtrField<std::string> protobufStrList(const std::vector<std::string>& v) {
+  Protobuf::RepeatedPtrField<std::string> res;
+  for (auto& field : v) {
+    *res.Add() = field;
+  }
+
+  return res;
+}
+
+class RouterTestStrictCheckOneHeader : public RouterTestBase,
+                                       public testing::WithParamInterface<std::string> {
+public:
+  RouterTestStrictCheckOneHeader() : RouterTestBase(false, false, protobufStrList({GetParam()})){};
+};
+
+INSTANTIATE_TEST_SUITE_P(StrictHeaderCheck, RouterTestStrictCheckOneHeader,
+                         testing::Values("x-envoy-upstream-rq-timeout-ms",
+                                         "x-envoy-upstream-rq-per-try-timeout-ms",
+                                         "x-envoy-max-retries", "x-envoy-retry-on",
+                                         "x-envoy-retry-grpc-on"));
+
+// Each test param instantiates a router that strict-checks one particular header.
+// This test decodes a set of headers with invalid values and asserts that the
+// strict header check only fails for the single header specified by the test param
+TEST_P(RouterTestStrictCheckOneHeader, SingleInvalidHeader) {
+  Http::TestHeaderMapImpl req_headers{
+      {"X-envoy-Upstream-rq-timeout-ms", "10.0"},
+      {"x-envoy-upstream-rq-per-try-timeout-ms", "1.0"},
+      {"x-envoy-max-retries", "2.0"},
+      {"x-envoy-retry-on", "5xx,cancelled"},            // 'cancelled' is an invalid entry
+      {"x-envoy-retry-grpc-on", "cancelled, internal"}, // spaces are considered errors
+  };
+  HttpTestUtility::addDefaultHeaders(req_headers);
+  auto checked_header = GetParam();
+
+  EXPECT_CALL(callbacks_.stream_info_,
+              setResponseFlag(StreamInfo::ResponseFlag::InvalidEnvoyRequestHeaders));
+
+  EXPECT_CALL(callbacks_, encodeHeaders_(_, _))
+      .WillOnce(Invoke([&](Http::HeaderMap& response_headers, bool end_stream) -> void {
+        EXPECT_EQ(enumToInt(Http::Code::BadRequest),
+                  Envoy::Http::Utility::getResponseStatus(response_headers));
+        EXPECT_FALSE(end_stream);
+      }));
+
+  EXPECT_CALL(callbacks_, encodeData(_, _))
+      .WillOnce(Invoke([&](Buffer::Instance& data, bool end_stream) -> void {
+        EXPECT_THAT(data.toString(),
+                    StartsWith(fmt::format("invalid header '{}' with value ", checked_header)));
+        EXPECT_TRUE(end_stream);
+      }));
+
+  EXPECT_EQ(Http::FilterHeadersStatus::StopIteration, router_.decodeHeaders(req_headers, true));
+  EXPECT_EQ(callbacks_.details_,
+            fmt::format("request_headers_failed_strict_check{{{}}}", checked_header));
+}
+
+class RouterTestStrictCheckSomeHeaders
+    : public RouterTestBase,
+      public testing::WithParamInterface<std::vector<std::string>> {
+public:
+  RouterTestStrictCheckSomeHeaders() : RouterTestBase(false, false, protobufStrList(GetParam())){};
+};
+
+INSTANTIATE_TEST_SUITE_P(StrictHeaderCheck, RouterTestStrictCheckSomeHeaders,
+                         testing::Values(std::vector<std::string>{"x-envoy-upstream-rq-timeout-ms",
+                                                                  "x-envoy-max-retries"},
+                                         std::vector<std::string>{}));
+
+// Request has headers with invalid values, but headers are *excluded* from the
+// set to which strict-checks apply. Assert that these headers are not rejected.
+TEST_P(RouterTestStrictCheckSomeHeaders, IgnoreOmittedHeaders) {
+  // Invalid, but excluded from the configured set of headers to strictly-check
+  Http::TestHeaderMapImpl headers{
+      {"x-envoy-upstream-rq-per-try-timeout-ms", "1.0"},
+      {"x-envoy-upstream-rq-timeout-ms", "5000"},
+      {"x-envoy-retry-on", "5xx,cancelled"},
+  };
+  HttpTestUtility::addDefaultHeaders(headers);
+
+  expectResponseTimerCreate();
+  EXPECT_EQ(Http::FilterHeadersStatus::StopIteration, router_.decodeHeaders(headers, true));
+  router_.onDestroy();
+}
+
+const std::vector<std::string> SUPPORTED_STRICT_CHECKED_HEADERS = {
+    "x-envoy-upstream-rq-timeout-ms", "x-envoy-upstream-rq-per-try-timeout-ms", "x-envoy-retry-on",
+    "x-envoy-retry-grpc-on", "x-envoy-max-retries"};
+
+class RouterTestStrictCheckAllHeaders
+    : public RouterTestBase,
+      public testing::WithParamInterface<std::tuple<std::string, std::string>> {
+public:
+  RouterTestStrictCheckAllHeaders()
+      : RouterTestBase(false, false, protobufStrList(SUPPORTED_STRICT_CHECKED_HEADERS)){};
+};
+
+INSTANTIATE_TEST_SUITE_P(StrictHeaderCheck, RouterTestStrictCheckAllHeaders,
+                         testing::Combine(testing::ValuesIn(SUPPORTED_STRICT_CHECKED_HEADERS),
+                                          testing::ValuesIn(SUPPORTED_STRICT_CHECKED_HEADERS)));
+
+// Each instance of this test configures a router to strict-validate all
+// supported headers and asserts that a request with invalid values set for some
+// *pair* of headers is rejected.
+TEST_P(RouterTestStrictCheckAllHeaders, MultipleInvalidHeaders) {
+  const auto& header1 = std::get<0>(GetParam());
+  const auto& header2 = std::get<1>(GetParam());
+  Http::TestHeaderMapImpl headers{{header1, "invalid"}, {header2, "invalid"}};
+  HttpTestUtility::addDefaultHeaders(headers);
+
+  EXPECT_CALL(callbacks_.stream_info_,
+              setResponseFlag(StreamInfo::ResponseFlag::InvalidEnvoyRequestHeaders));
+
+  EXPECT_CALL(callbacks_, encodeHeaders_(_, _))
+      .WillOnce(Invoke([&](Http::HeaderMap& response_headers, bool end_stream) -> void {
+        EXPECT_EQ(enumToInt(Http::Code::BadRequest),
+                  Envoy::Http::Utility::getResponseStatus(response_headers));
+        EXPECT_FALSE(end_stream);
+      }));
+
+  EXPECT_EQ(Http::FilterHeadersStatus::StopIteration, router_.decodeHeaders(headers, true));
+  EXPECT_THAT(callbacks_.details_,
+              StartsWith(fmt::format("request_headers_failed_strict_check{{")));
+  router_.onDestroy();
+}
+
+// Request has headers with invalid values, but headers are *excluded* from the
+// set to which strict-checks apply. Assert that these headers are not rejected.
+TEST(RouterFilterUtilityTest, StrictCheckValidHeaders) {
+  Http::TestHeaderMapImpl headers{
+      {"X-envoy-Upstream-rq-timeout-ms", "100"},
+      {"x-envoy-upstream-rq-per-try-timeout-ms", "100"},
+      {"x-envoy-max-retries", "2"},
+      {"not-checked", "always passes"},
+      {"x-envoy-retry-on",
+       "5xx,gateway-error,retriable-4xx,refused-stream,connect-failure,retriable-status-codes"},
+      {"x-envoy-retry-grpc-on",
+       "cancelled,internal,deadline-exceeded,resource-exhausted,unavailable"},
+  };
+
+  for (const auto& target : SUPPORTED_STRICT_CHECKED_HEADERS) {
+    EXPECT_TRUE(
+        FilterUtility::StrictHeaderChecker::checkHeader(headers, Http::LowerCaseString(target))
+            .valid_)
+        << fmt::format("'{}' should have passed strict validation", target);
+  }
+
+  Http::TestHeaderMapImpl failing_headers{
+      {"X-envoy-Upstream-rq-timeout-ms", "10.0"},
+      {"x-envoy-upstream-rq-per-try-timeout-ms", "1.0"},
+      {"x-envoy-max-retries", "2.0"},
+      {"x-envoy-retry-on", "5xx,cancelled"},            // 'cancelled' is an invalid entry
+      {"x-envoy-retry-grpc-on", "cancelled, internal"}, // spaces are considered errors
+  };
+
+  for (const auto& target : SUPPORTED_STRICT_CHECKED_HEADERS) {
+    EXPECT_FALSE(FilterUtility::StrictHeaderChecker::checkHeader(failing_headers,
+                                                                 Http::LowerCaseString(target))
+                     .valid_)
+        << fmt::format("'{}' should have failed strict validation", target);
+  }
 }
 
 } // namespace Router