+++ test/extensions/filters/http/router/config_test.cc
@@ -47,6 +47,30 @@ TEST(RouterFilterConfigTest, BadRouterFilterConfig) {
   EXPECT_THROW(factory.createFilterFactory(*json_config, "stats", context), Json::Exception);
 }
 
+TEST(RouterFilterConfigTest, RouterFilterWithUnsupportedStrictHeaderCheck) {
+  const std::string yaml = R"EOF(
+  strict_check_headers:
+  - unsupportedHeader
+  )EOF";
+
+  envoy::config::filter::http::router::v2::Router router_config;
+  TestUtility::loadFromYaml(yaml, router_config);
+
+  NiceMock<Server::Configuration::MockFactoryContext> context;
+  RouterFilterConfig factory;
+  EXPECT_THROW_WITH_MESSAGE(
+      factory.createFilterFactoryFromProto(router_config, "stats", context),
+      ProtoValidationException,
+      "Proto constraint validation failed (RouterValidationError.StrictCheckHeaders[i]: "
+      "[\"value must be in list \" ["
+      "\"x-envoy-upstream-rq-timeout-ms\" "
+      "\"x-envoy-upstream-rq-per-try-timeout-ms\" "
+      "\"x-envoy-max-retries\" "
+      "\"x-envoy-retry-grpc-on\" "
+      "\"x-envoy-retry-on\""
+      "]]): strict_check_headers: \"unsupportedHeader\"\n");
+}
+
 TEST(RouterFilterConfigTest, RouterV2Filter) {
   envoy::config::filter::http::router::v2::Router router_config;
   router_config.mutable_dynamic_stats()->set_value(true);